4.0.0 Stability Bug fixes, style fixes, optimizations, enhancements [4657] Reactify Global App (Users) - craftercms/craftercms#4657
