+++ ui/app/src/components/EditUserDialog/index.ts
@@ -14,8 +14,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default } from './UserEditDialog';
+export { default } from './EditUserDialog';
 
-export * from './UserEditDialog';
-export * from './styles';
-export * from './UserInfoDialogUI';
+export * from './EditUserDialog';
+
+export * from './EditUserDialogUI';
+
+export * from './EditUserDialogContainer';
